Problems abound -- broken tests
The McCode-conductor broke due to missing `--ncount` in the splitrun
parser. After adding that back-in, it became apparent that tests have
not been run in some time as the same issue should have been caught by a
test in this repository.

Running the tests now raises a number of issues. The main  splitrun test
has an unresolved issue which seems to be due to the use of guides in
the test instrument.

Other broken tests remain.